requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
cors==1.0.1
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.1
future==1.0.0
gevent==25.9.1
greenlet==3.3.0
idna==3.11
psycopg2==2.9.11
PyJWT==2.10.1
PySocks==1.7.1
requests==2.32.5
requests-file==3.0.1
sqlparse==0.5.5
tldextract==5.3.0
urllib3==2.6.2
zope.event==6.1
zope.interface==8.1.1
